requirements.txt
Flask
waitress
pandas
tensorflow
numpy
matplotlib
yfinance
keras
python-dateutil
scikit-learn
XlsxWriter